Specify useGoStep: false for tmsUpload step
The Golang implementation of the tmsUpload step is now a default one. To run the tests against the Groovy implementation of the step (as it was done previously) that is still available and can be used by Customers, the useGoStep: false parameter has to be explicitly provided for the tmsUpload step.